Don't implicitly set path in CSV reader functions
The CSV reading functions now take a raw path, rather than a filename in the
vc2_data_tables/csv directory. This allows these functions to be used
externally. It also makes it possible to put the test CSV files in the test
directory so they don't have to live alongside the 'real' CSV files.